Make generic FilterPanel and FilterSidebar classes for use with other tools
