+++ src/main/kotlin/io/github/inductiveautomation/kindling/log/LogPanel.kt
@@ -1,9 +1,10 @@
 package io.github.inductiveautomation.kindling.log
 
-import com.formdev.flatlaf.extras.components.FlatTabbedPane
 import com.formdev.flatlaf.ui.FlatScrollBarUI
 import io.github.inductiveautomation.kindling.core.Detail.BodyLine
 import io.github.inductiveautomation.kindling.core.DetailsPane
+import io.github.inductiveautomation.kindling.core.Filter
+import io.github.inductiveautomation.kindling.core.FilterSidebar
 import io.github.inductiveautomation.kindling.core.Kindling.Preferences.Advanced.Debug
 import io.github.inductiveautomation.kindling.core.Kindling.Preferences.Advanced.HyperlinkStrategy
 import io.github.inductiveautomation.kindling.core.Kindling.Preferences.General.ShowFullLoggerNames
@@ -14,7 +15,6 @@ import io.github.inductiveautomation.kindling.core.ToolPanel
 import io.github.inductiveautomation.kindling.log.LogViewer.ShowDensity
 import io.github.inductiveautomation.kindling.log.LogViewer.TimeStampFormatter
 import io.github.inductiveautomation.kindling.utils.Action
-import io.github.inductiveautomation.kindling.utils.Column
 import io.github.inductiveautomation.kindling.utils.EDT_SCOPE
 import io.github.inductiveautomation.kindling.utils.FlatScrollPane
 import io.github.inductiveautomation.kindling.utils.MajorVersion
@@ -38,7 +38,6 @@ import java.awt.RenderingHints
 import java.awt.geom.AffineTransform
 import java.time.Duration
 import java.time.Instant
-import java.util.EventListener
 import java.util.Vector
 import javax.swing.Icon
 import javax.swing.JCheckBox
@@ -95,7 +94,26 @@ class LogPanel(
         verticalScrollBar = densityDisplay
     }
 
-    private val sidebar = Sidebar(rawData)
+    private val sidebar = FilterSidebar(
+        NamePanel(rawData),
+        LevelPanel(rawData),
+        if (rawData.first() is SystemLogEvent) {
+            @Suppress("UNCHECKED_CAST")
+            MDCPanel(rawData as List<SystemLogEvent>)
+        } else {
+            null
+        },
+        if (rawData.first() is SystemLogEvent) {
+            @Suppress("UNCHECKED_CAST")
+            ThreadPanel(rawData as List<SystemLogEvent>)
+        } else {
+            null
+        },
+        TimePanel(
+            lowerBound = rawData.first().timestamp,
+            upperBound = rawData.last().timestamp,
+        ),
+    )
 
     private val details = DetailsPane()
 
@@ -414,90 +432,6 @@ class LogPanel(
         }
     }
 
-    private class Sidebar(rawData: List<LogEvent>) : FlatTabbedPane() {
-        private val names = NamePanel(rawData)
-        private val levels = LevelPanel(rawData)
-
-        private val systemLogs: Boolean = rawData.first() is SystemLogEvent
-
-        private val mdc: MDCPanel? = if (systemLogs) {
-            @Suppress("UNCHECKED_CAST")
-            MDCPanel(rawData as List<SystemLogEvent>)
-        } else {
-            null
-        }
-
-        private val threads: ThreadPanel? = if (systemLogs) {
-            @Suppress("UNCHECKED_CAST")
-            ThreadPanel(rawData as List<SystemLogEvent>)
-        } else {
-            null
-        }
-
-        private val time = TimePanel(
-            lowerBound = rawData.first().timestamp,
-            upperBound = rawData.last().timestamp,
-        )
-
-        val filterPanels: List<LogFilterPanel> = listOfNotNull(
-            names,
-            levels,
-            time,
-            mdc,
-            threads,
-        )
-
-        init {
-            tabLayoutPolicy = SCROLL_TAB_LAYOUT
-            tabsPopupPolicy = TabsPopupPolicy.asNeeded
-            scrollButtonsPolicy = ScrollButtonsPolicy.never
-            tabWidthMode = TabWidthMode.equal
-            tabType = TabType.underlined
-            tabHeight = 16
-
-            for (i in filterPanels.indices) {
-                val filterPanel = filterPanels[i]
-                addTab(filterPanel.tabName, filterPanel.component)
-
-                filterPanel.addFilterChangeListener {
-                    filterPanel.updateTabState()
-                    selectedIndex = i
-                }
-            }
-
-            attachPopupMenu { event ->
-                val tabIndex = indexAtLocation(event.x, event.y)
-                if (tabIndex == -1) return@attachPopupMenu null
-
-                JPopupMenu().apply {
-                    add(
-                        Action("Reset") {
-                            filterPanels[tabIndex].reset()
-                        },
-                    )
-                }
-            }
-        }
-
-        private fun LogFilterPanel.updateTabState() {
-            val index = indexOfComponent(component)
-            if (isFilterApplied()) {
-                setBackgroundAt(index, UIManager.getColor("TabbedPane.focusColor"))
-                setTitleAt(index, "$tabName *")
-            } else {
-                setBackgroundAt(index, UIManager.getColor("TabbedPane.background"))
-                setTitleAt(index, tabName)
-            }
-        }
-
-        override fun updateUI() {
-            super.updateUI()
-            @Suppress("UNNECESSARY_SAFE_CALL")
-            filterPanels?.forEach {
-                it.updateTabState()
-            }
-        }
-    }
 
     companion object {
         private val BACKGROUND = CoroutineScope(Dispatchers.Default)
@@ -524,28 +458,4 @@ class LogPanel(
     }
 }
 
-fun interface LogFilter {
-    /**
-     * Return true if this filter should display this event.
-     */
-    fun filter(event: LogEvent): Boolean
-}
-
-fun interface FilterChangeListener : EventListener {
-    fun filterChanged()
-}
-
-interface LogFilterPanel : LogFilter {
-    val tabName: String
-    fun isFilterApplied(): Boolean
-    val component: JComponent
-    fun addFilterChangeListener(listener: FilterChangeListener)
-
-    fun reset()
-
-    fun customizePopupMenu(
-        menu: JPopupMenu,
-        column: Column<out LogEvent, *>,
-        event: LogEvent,
-    ) = Unit
-}
+typealias LogFilter = Filter<LogEvent>
